apps/accounts/auth_serializers.py
from rest_framework import serializers
from rest_framework_simplejwt.serializers import TokenObtainPairSerializer
from rest_framework_simplejwt.tokens import UntypedToken

from .serializers import SigninUserSerializer


class CustomTokenObtainPairSerializer(TokenObtainPairSerializer):

    def validate(self, attrs):
        data = super().validate(attrs)
        refresh = self.get_token(self.user)
        data['refresh'] = str(refresh)
        data['access'] = str(refresh.access_token)
        data.pop('refresh', data)
        # extra responses
        user_serialized = SigninUserSerializer(self.user)
        data['user'] = user_serialized.data

        return data
